Добавлен AsyncTaskLoader вместо AsyncTask

+++ app/src/main/java/ru/ipimenov/violetrainbow/CatalogFragment.java
@@ -3,11 +3,10 @@ package ru.ipimenov.violetrainbow;
 import android.annotation.SuppressLint;
 import android.content.Intent;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Toast;
+import android.widget.ProgressBar;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -15,11 +14,15 @@ import androidx.fragment.app.Fragment;
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProviders;
+import androidx.loader.app.LoaderManager;
+import androidx.loader.content.Loader;
 import androidx.recyclerview.widget.GridLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
+import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 import ru.ipimenov.violetrainbow.data.ForeignViolet;
 import ru.ipimenov.violetrainbow.data.MainViewModel;
@@ -28,7 +31,7 @@ import ru.ipimenov.violetrainbow.data.Violet;
 import ru.ipimenov.violetrainbow.utils.ContentUtils;
 import ru.ipimenov.violetrainbow.utils.NetworkUtils;
 
-public class CatalogFragment extends Fragment {
+public class CatalogFragment extends Fragment implements LoaderManager.LoaderCallbacks<String> {
 
     private int catalog;
     private int page;
@@ -36,6 +39,12 @@ public class CatalogFragment extends Fragment {
     private MainViewModel viewModel;
     private VioletAdapter violetAdapter;
     private ArrayList<Violet> violets;
+
+    private static final int LOADER_ID = 133;
+    private LoaderManager loaderManager;
+
+    private ProgressBar progressBarLoading;
+    private static boolean isLoading = false;
 
     CatalogFragment(int catalog, int page) {
         this.catalog = catalog;
@@ -46,6 +55,9 @@ public class CatalogFragment extends Fragment {
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container,
                              @Nullable Bundle savedInstanceState) {
+        progressBarLoading = MainActivity.progressBarLoading;
+
+        loaderManager = LoaderManager.getInstance(this);
 
         viewModel = ViewModelProviders.of(this).get(MainViewModel.class);
 
@@ -55,12 +67,19 @@ public class CatalogFragment extends Fragment {
 
         violetAdapter = new VioletAdapter();
 
-        String content = NetworkUtils.getContentFromNetwork(catalog, page);
-        violets = ContentUtils.getVioletsFromContent(content);
+        downloadData();
 
-        if (violets != null && !violets.isEmpty()) {
-            setVioletsToDatabase();
-        }
+//        URL url = NetworkUtils.buildURL(catalog, page);
+//        Bundle bundle = new Bundle();
+//        bundle.putString("url", url.toString());
+//        loaderManager.restartLoader(LOADER_ID, bundle, this);
+
+//        String content = NetworkUtils.getContentFromNetwork(catalog, page);
+//        violets = ContentUtils.getVioletsFromContent(content);
+
+//        if (violets != null && !violets.isEmpty()) {
+//            setVioletsToDatabase();
+//        }
         getVioletsFromDatabase();
 
         recyclerView.setAdapter(violetAdapter);
@@ -79,7 +98,10 @@ public class CatalogFragment extends Fragment {
         violetAdapter.setOnReachEndListener(new VioletAdapter.OnReachEndListener() {
             @Override
             public void onReachEnd() {
-                Toast.makeText(getActivity(), "Конец списка", Toast.LENGTH_SHORT).show();
+                if (!isLoading) {
+                    downloadData();
+//                    Toast.makeText(getActivity(), "Конец списка", Toast.LENGTH_SHORT).show();
+                }
             }
         });
 
@@ -89,22 +111,37 @@ public class CatalogFragment extends Fragment {
     private void setVioletsToDatabase() {
         switch (catalog) {
             case NetworkUtils.RU_UA_SELECTION:
-                viewModel.deleteAllRuUaViolets();
+                if (page == 1) {
+                    viewModel.deleteAllRuUaViolets();
+                    violetAdapter.clear();
+                }
                 for (Violet violet : violets) {
                     viewModel.insertRuUaViolet(violet);
                 }
+                violetAdapter.addViolets(violets);
+                page++;
                 break;
             case NetworkUtils.FOREIGN_SELECTION:
-                viewModel.deleteAllForeignViolets();
+                if (page == 1) {
+                    viewModel.deleteAllForeignViolets();
+                    violetAdapter.clear();
+                }
                 for (Violet violet : violets) {
                     viewModel.insertForeignViolet(new ForeignViolet(violet));
                 }
+                violetAdapter.addViolets(violets);
+                page++;
                 break;
             case NetworkUtils.MINI_SELECTION:
-                viewModel.deleteAllMiniViolets();
+                if (page == 1) {
+                    viewModel.deleteAllMiniViolets();
+                    violetAdapter.clear();
+                }
                 for (Violet violet : violets) {
                     viewModel.insertMiniViolet(new MiniViolet(violet));
                 }
+                violetAdapter.addViolets(violets);
+                page++;
                 break;
         }
     }
@@ -117,7 +154,9 @@ public class CatalogFragment extends Fragment {
                 violetsFromLiveDataR.observe(this, new Observer<List<Violet>>() {
                     @Override
                     public void onChanged(List<Violet> violets) {
-                        violetAdapter.setViolets(violets);
+                        if (page == 1) {
+                            violetAdapter.setViolets(violets);
+                        }
                     }
                 });
                 break;
@@ -127,7 +166,9 @@ public class CatalogFragment extends Fragment {
                     @Override
                     public void onChanged(List<ForeignViolet> violets) {
                         List<Violet> violetsF = new ArrayList<Violet>(violets);
-                        violetAdapter.setViolets(violetsF);
+                        if (page == 1) {
+                            violetAdapter.setViolets(violetsF);
+                        }
                     }
                 });
                 break;
@@ -137,10 +178,50 @@ public class CatalogFragment extends Fragment {
                     @Override
                     public void onChanged(List<MiniViolet> violets) {
                         List<Violet> violetsM = new ArrayList<Violet>(violets);
-                        violetAdapter.setViolets(violetsM);
+                        if (page == 1) {
+                            violetAdapter.setViolets(violetsM);
+                        }
                     }
                 });
                 break;
         }
+    }
+
+    private void downloadData() {
+        URL url = NetworkUtils.buildURL(catalog, page);
+        Bundle bundle = new Bundle();
+        bundle.putString("url", url.toString());
+        loaderManager.restartLoader(catalog, bundle, this);
+    }
+
+    @NonNull
+    @Override
+    public Loader<String> onCreateLoader(int id, @Nullable Bundle args) {
+        NetworkUtils.ContentLoader contentLoader = new NetworkUtils.ContentLoader(Objects.requireNonNull(getContext()), args);
+//        NetworkUtils.ContentLoader contentLoader = new NetworkUtils.ContentLoader(LOADER_ID, args);
+        contentLoader.setOnStartLoadingListener(new NetworkUtils.ContentLoader.OnStartLoadingListener() {
+            @Override
+            public void onStartLoading() {
+                progressBarLoading.setVisibility(View.VISIBLE);
+                isLoading = true;
+            }
+        });
+        return contentLoader;
+    }
+
+    @Override
+    public void onLoadFinished(@NonNull Loader<String> loader, String data) {
+        violets = ContentUtils.getVioletsFromContent(data);
+        if (violets != null && !violets.isEmpty()) {
+            setVioletsToDatabase();
+        }
+        progressBarLoading.setVisibility(View.INVISIBLE);
+        isLoading = false;
+        loaderManager.destroyLoader(catalog);
+    }
+
+    @Override
+    public void onLoaderReset(@NonNull Loader<String> loader) {
+
     }
 }
